More click intersect work p much works but you have to be right next to the model and it behaves p erratically

+++ src/Vertex.h
@@ -1,8 +1,8 @@
 class Vertex
 {
 public:
-	float x;
-	float y;
-	float z;
-	float w;
+	GLdouble x;
+	GLdouble y;
+	GLdouble z;
+	GLdouble w;
 };
